configure capybara to take screenshots of errors
Enable capybara-screenshot to post to S3 in Travis. Secure credentials
given that can only create files in a specific bucket for test
screenies.

* Provides ENV var for configuring capybara-screenshot to save to S3,
  upgrade aws-sdk so that it can.
* Upgrades paperclip because of the upgrade to aws-sdk.
* some formatting tweaks to Travis config that Travis CLI really wanted

* Include S3_REGION, now required by aws-sdk and api v2

add an abort during initialization if only a partial S3 config is
supplied

With the upgrade of paperclip and the aws-sdk, S3_REGION is now required
else calls can't be made to the S3 API. Because there are existing
installs that probably lack this environment variable until a human can
add it to the configuration, we feel it is better to abort starting the
web app than to start and store cookbooks on disk unbeknownst to the
admins and user.

Moved the logic of whether Supermarket is configured for S3 to a library
class to make testing the cases clearer. Initializers are wicked and
unforgiving; keep them simple.

* make the check for S3_PRIVATE_OBJECTS rigorous

Because the check for .present? on the environment variable could cause
unexpected behavior if someone sets it to "false".

    > export S3_PRIVATE_OBJECTS="false"
    > be rails c
    Loading development environment (Rails 4.2.7.1)
    [1] pry(main)> ENV['S3_PRIVATE_OBJECTS']
    => "false"
    [2] pry(main)> ENV['S3_PRIVATE_OBJECTS'].present?
    => true

* replace cookbook upload api spec with tests for upload validations

Our API spec helpers tar-and-gzip given fixtures under the hood. Instead
of overhauling that to just test that an error from Paperclip floats up
through the API response (already well tested by other specs), test the
validation expectation on the model.

Signed-off-by: Robb Kidd <rkidd@chef.io>