make navmenu hide/show consistent between top nav and drawer nav
The menu links in the top nav and in the left-side drawer should hide
and appear at the same size. Previously, there was a range of browser
widths ("small" to "medium") that would make the top nav disappear
before the drawer nav would appear.

* move create account/sign in links to user menu

Because the "Create Account" and "Sign In" links are about user
management, it makes sense for them to be in the usermenu div that has
display logic to ensure that it is always present at the top of the
page. At widths where the mainav and usermenu collide (medium and
smaller), the mainnav div is told to disappear from the top and appear
in the hide-away menu on the left.

* Reduced the margin between the links to help get the whole mainnav menu
  width to a size below the collision at medium width.
* Removed the outside margins from the first and last menu links to help
  with the collision widths. The margins there were unnecessary because
  they were not separating links from each other and the divs have the
  spacing they need already.

This also moves the hide/show declaration into a foundation class on the
menu element itself. Nell and I concluded that it would be clearer for
the time being to have the foundation show/hide declaration in the
markup instead of declaring show in markup in one place and hide in SCSS
in another place.

Also also, because top mainnav hides at medium and smaller, we've set
the test browser width to be firmly larger than the medium range so that
the menu we expect most human users to interact with is available on the
page.  Medium's current upper bound is 1025 which puts capybara's
default of 1024 in medium resulting in the mainnav links appearing in
the left-off-canvas-menu. Instead of converting all feature specs over
to going through clicking on the hamburger menu (left-off-canvas-toggle)
and then using the links that appear there, I've opted to set the test
browser to be larger-than-medium so that the mainnav menu appears at the
top in the test environment.

Signed-off-by: Robb Kidd <rkidd@chef.io>